Added a whole new serialization system.
I added our new serialization system. I might replace the old one with
this one, as this one supports different data-types. If the data-type
is Integer, it looks for integers, and if the data-type is String it's
looking for a String. Pretty self-explanitory. There might be bugs to
fix though.